chore: address issue #147
